controller ok

+++ src/main/java/com/flow/api/account/BankAccountService.java
@@ -2,9 +2,11 @@ package com.flow.api.account;
 
 import com.flow.api.account.dto.BankAccountResponse;
 import com.flow.api.account.dto.CreateBankAccountRequest;
+import com.flow.api.account.dto.UpdateBankAccountRequest;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,28 +25,76 @@ public class BankAccountService {
         return bankAccountRepository.save(bankAccount);
     }
 
-   public Optional<BankAccount> findBankAccountById(Long id) {
-        return bankAccountRepository.findById(id);
+    public Optional<BankAccountResponse> findBankAccountById(Long id) {
+
+        Optional<BankAccount> optionalBankAccount= bankAccountRepository.findById(id);
+
+        if(optionalBankAccount.isPresent()){
+
+            BankAccount foundBankAccount = optionalBankAccount.get();
+
+            BankAccountResponse response = new BankAccountResponse(
+                    foundBankAccount.getId(),
+                    foundBankAccount.getName(),
+                    foundBankAccount.getInitialBalance(),
+                    foundBankAccount.getCreationDate()
+            );
+
+            return Optional.of(response);
+
+        } else {
+                return Optional.empty();
+
+        }
+
    }
 
-   public List<BankAccount> findAllBankAccount(){
-        return bankAccountRepository.findAll();
-   }
+    public List<BankAccountResponse> findAllBankAccounts() {
 
-  public void deleteBankAccountById(Long id) {
-        bankAccountRepository.deleteById(id);
+        List<BankAccount> bankAccounts =
+                bankAccountRepository.findAll();
+
+        List<BankAccountResponse> responses =
+                new ArrayList<>();
+        for (BankAccount bankAccount : bankAccounts) {
+            BankAccountResponse response = new BankAccountResponse(
+                    bankAccount.getId(),
+                    bankAccount.getName(),
+                    bankAccount.getInitialBalance(),
+                    bankAccount.getCreationDate()
+            );
+            responses.add(response);
+        }
+
+        return responses;
     }
 
-  public Optional<BankAccount> updateBankAccount(Long id, String name, BigDecimal initialBalance) {
+  public boolean deleteBankAccountById(Long id) {
+
+        if(bankAccountRepository.existsById(id)){
+            bankAccountRepository.deleteById(id);
+            return true;
+        }
+
+        return false;
+  }
+
+  public Optional<BankAccountResponse> updateBankAccount(Long id, UpdateBankAccountRequest request) {
         Optional<BankAccount> optionalBankAccount = bankAccountRepository.findById(id);
 
       if (optionalBankAccount.isPresent()) {
             BankAccount bankAccount = optionalBankAccount.get();
-            bankAccount.setName(name);
-            bankAccount.setInitialBalance(initialBalance);
+            bankAccount.setName(request.name());
+            bankAccount.setInitialBalance(request.initialBalance());
             bankAccountRepository.save(bankAccount);
 
-            return Optional.of(bankAccount);
+            BankAccountResponse response = new BankAccountResponse(
+                    bankAccount.getId(),
+                    bankAccount.getName(),
+                    bankAccount.getInitialBalance(),
+                    bankAccount.getCreationDate()
+            );
+            return Optional.of(response);
 
         } else {
             return Optional.empty();}
@@ -67,4 +117,5 @@ public class BankAccountService {
       );
 
   }
+
 }
